v1.3 - Quality of Life updates
Lower AxisLock threshold so the mouse doesn't travel as far to move the pen.
Compensate for zoom only when zoomed in
When compensating for zoom, restrict to a delta of 1 to keep the pen close to where the mouse moves.
Fix bug where Alt-Tab to switch to another window will leave us stuck in snapToGrid mode. (Made frame a member variable, so we can reference it in creating the WindowFocusListener to accomplish this bug fix)
Update help screen to reflect v1.3 (forgot to do that for v1.2)